리액트 학습
- useRef의 값은 변경되도 컴포넌트 렌더링에 영향을 끼치지 않는다.
- 반대로 컴포넌트가 렌더링 되더라도 useRef 값은 영향을 받지 않는다.
