+++ React/README.md
@@ -214,3 +214,8 @@ return (
         </div>
     );
     ```
+
+## useRef로 컴포넌트 안의 변수 만들기
+- 값을 바꿔도 재렌더링되지 않았으면 좋겠는 값을 다룰 때 사용하면 좋다.
+- setTimeout, setInterval, 서드파티 라이버르리를 사용하기 위해 사용된 인스턴스 등을 관리할 때 사용
+- 중요한 건 useRef 값은 변경되도 재렌더링되지 않는다는 것!
