+++ React/react-tutorial/begin-react/src/App.js
@@ -1,9 +1,34 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import UserList from "./UserList";
 
 function App() {
+  const users = [
+    {
+      id: 1,
+      username: "poppo",
+      email: "poppo@gmail.com"
+    },
+    {
+      id: 2,
+      username: "saul",
+      email: "saul@gmail.com"
+    },
+    {
+      id: 3,
+      username: "ita",
+      email: "ita@naver.com"
+    },
+  ];
+
+  const nextId = useRef(4);
+
+  const onCreate = () => {
+    console.log(nextId.current);
+    nextId.current += 1;
+  };
+
   return (
-      <UserList/>
+      <UserList users={users}/>
   );
 }
